scripts/schedule-vision/requirements.txt
openpyxl>=3.1.0
Pillow>=10.0
scikit-learn>=1.3
numpy>=1.26
joblib>=1.3

# VLM (optional — only needed for discover.py)
# mlx>=0.22
# mlx-vlm>=0.1
# huggingface-hub>=0.20
